Allow Sqitch/QGIS prep to run on external host. No expectation of postgres db on external host. Restructure logic, add tests

+++ docker/db.py
@@ -42,7 +42,13 @@ def connection_string(admin=False):
     pg_db = pg_details['pg_db']
 
     if admin:
-        db_name = 'postgres'
+        if pg_host == 'localhost':
+            db_name = 'postgres'
+        else:
+            # External databases only use admin connection for version check.
+            # Can connect to main data DB for this, `postgres` db not required.
+            # Should allow connection even when `postgres` database does not exist.
+            db_name = pg_db
     else:
         db_name = pg_db
 
@@ -156,9 +162,10 @@ def pg_isready():
     """
     try:
         result = pg_version_check()
-    except:
-        err_msg = f'Error checking version. Ensure Postgres connection details are correct.'
-        logging.getLogger('pgosm-flex').error(err_msg)
+    except AttributeError:
+        err_msg = 'Error checking version, likely waiting for Postgres to start.'
+        err_msg += ' Only an error if it does not go away after a few attempts.'
+        logging.getLogger('pgosm-flex').warning(err_msg)
         return False
 
     if result is None:
@@ -169,24 +176,21 @@ def pg_isready():
 def prepare_pgosm_db(data_only, db_path):
     """Runs through series of steps to prepare database for PgOSM.
 
-    Only should run on in-Docker database, intentionally leaving sub-components
-    hard coded to `pgosm` database.
-
     Parameters
     --------------------------
     data_only : bool
     db_path : str
     """
-    # Outer logic should skip this step. Additional check here is safety to ensure
-    # this does not run outside Docker image
-    if not pg_conn_parts()['pg_host'] == 'localhost':
-        LOGGER.error('Attempted running db.prepare_pgosm_db() on non-Docker database.')
-        return False
 
-    drop_pgosm_db()
-    create_pgosm_db()
+    if pg_conn_parts()['pg_host'] == 'localhost':
+        LOGGER.debug('Running standard database prep for in-Docker operation. Includes DROP/CREATE DATABASE')
+        drop_pgosm_db()
+        create_pgosm_db()
+    else:
+        LOGGER.info('Using external database. Ensure the target database is setup properly for PgOSM Flex with PostGIS, osm schema, and proper permissions.')
+
     if not data_only:
-        LOGGER.info('Loading extras via Sqitch.')
+        LOGGER.info('Loading extras via Sqitch plus QGIS styles.')
         run_sqitch_prep(db_path)
         load_qgis_styles(db_path)
     else:
@@ -194,31 +198,32 @@ def prepare_pgosm_db(data_only, db_path):
 
 
 def pg_version_check():
-    """Checks Postgres version.
+    """Checks Postgres machine-readible server_version_num.
 
     Sends to logs and returns value.
 
     Results
     --------------------
-    pg_version : str
+    pg_version : int
     """
     sql_raw = """
 SELECT setting
     FROM pg_catalog.pg_settings
     WHERE name = 'server_version_num'
-;
-"""
+;"""
 
     with get_db_conn(conn_string=os.environ['PGOSM_CONN_PG']) as conn:
         cur = conn.cursor()
         cur.execute(sql_raw)
         results = cur.fetchone()
 
-    pg_version = results[0]
-    if pg_version < '120000':
-        raise EnvironmentError('Postgres version {pg_verson} not supported. Postgres 12+ required.')
+    # It's an int https://www.postgresql.org/docs/current/runtime-config-preset.html#GUC-SERVER-VERSION-NUM
+    pg_version = int(results[0])
+    if pg_version < 120000:
+        err_msg = f'Postgres version {pg_version} not supported. Postgres 12+ required.'
+        LOGGER.error(err_msg)
+        sys.exit(9)
 
-    LOGGER.debug(f'Postgres version number {pg_version}')
     return pg_version
 
 
@@ -227,6 +232,10 @@ def drop_pgosm_db():
 
     Intentionally hard coded to `pgosm` database for in-Docker use only.
     """
+    if not pg_conn_parts()['pg_host'] == 'localhost':
+        LOGGER.error('Attempted to drop database external from Docker. Not doing that')
+        return False
+
     sql_raw = 'DROP DATABASE IF EXISTS pgosm;'
     conn = get_db_conn(conn_string=os.environ['PGOSM_CONN_PG'])
 
@@ -242,6 +251,10 @@ def create_pgosm_db():
 
     Intentionally hard coded to `pgosm` database for in-Docker use only.
     """
+    if not pg_conn_parts()['pg_host'] == 'localhost':
+        LOGGER.error('Attempted to create database external from Docker. Not doing that')
+        return False
+
     sql_raw = 'CREATE DATABASE pgosm;'
     conn = get_db_conn(conn_string=os.environ['PGOSM_CONN_PG'])
